+++ photon-model-adapters/gcp/src/test/java/com/vmware/photon/controller/model/adapters/gcp/GCPTestUtil.java
@@ -97,8 +97,8 @@ public class GCPTestUtil {
             ".com/compute/v1/projects/%s/zones/%s/diskTypes/pd-standard";
     private static final long ONE_HOUR_DIFFERENCE_MICROS = TimeUnit.HOURS.toMicros(1);
     private static final long WAIT_INTERVAL = 1000;
-    private static final int MIN_CPU_COUNT = 1;
-    private static final int MIN_MEMORY_BYTES = 1024;
+    private static final long MIN_CPU_COUNT = 1;
+    private static final long MIN_MEMORY_BYTES = 1024;
     private static final JsonFactory JSON_FACTORY = JacksonFactory.getDefaultInstance();
 
     /**