+++ photon-model-adapters/vsphere/adapter/src/test/java/com/vmware/photon/controller/model/adapters/vsphere/BaseVSphereAdapterTest.java
@@ -181,8 +181,8 @@ public class BaseVSphereAdapterTest {
         inPool.name = "resourcePool-" + UUID.randomUUID().toString();
         inPool.id = inPool.name;
 
-        inPool.minCpuCount = 1;
-        inPool.minMemoryBytes = 1024;
+        inPool.minCpuCount = 1L;
+        inPool.minMemoryBytes = 1024L;
 
         ResourcePoolState returnPool =
                 TestUtils.doPost(this.host, inPool, ResourcePoolState.class,