+++ photon-model-tasks/src/test/java/com/vmware/photon/controller/model/tasks/ModelUtils.java
@@ -108,8 +108,8 @@ public class ModelUtils {
         poolState.name = UUID.randomUUID().toString();
         poolState.id = poolState.name;
         poolState.documentSelfLink = poolState.id;
-        poolState.maxCpuCount = 1600;
-        poolState.minCpuCount = 16;
+        poolState.maxCpuCount = 1600L;
+        poolState.minCpuCount = 16L;
         poolState.currencyUnit = "Bitcoin";
         poolState.maxCpuCostPerMinute = 1.0;
         poolState.maxDiskCostPerMinute = 1.0;